pyproject.toml
[tool.poetry]
name = "text-to-sparql"
version = "0.1.0"
description = ""
authors = ["Dimitar Pavlovski <dimitar.pavlovski12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
langchain = "^0.3.25"
rmrkl = "^0.0.3"
prompts = "^0.0.1"
langchain-community = "^0.3.23"
langchain-ollama = "^0.3.2"
pydantic-ai = "^0.2.0"
sparqlwrapper = "^2.0.0"
datasets = "^3.6.0"
pymongo = "^4.12.1"
scikit-learn = "^1.6.1"
matplotlib = "^3.10.1"
pywikibot = "^10.0.0"
transformers = "4.52.1"
torch = "^2.7.0"
wikidata = "==0.8.1"
ujson = "^5.10.0"
poetry-core = "^2.1.3"
langchain-openai = "^0.3.16"
qdrant-client = "^1.15.1"
beautifulsoup4 = "^4.13.4"
bs4 = "^0.0.2"
langgraph = "^0.4.7"
ipython = "^9.3.0"
qwikidata = "^0.4.2"
deep-translator = "^1.11.4"
notebook = "^7.4.5"
sentence-transformers = "^5.1.1"
sentencepiece = "^0.2.1"
sacremoses = "^0.1.1"
pydantic = "^2.11.9"
dependency-injector = "^4.48.2"
streamlit = "^1.51.0"
plotly = "^6.5.0"
langchain-google-genai = "^2.0.0"
pandas-stubs = "~=2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
